// src/web3/types/chainInfo.ts
//-----------------------------------------------------------------------------------------------//
type chainInfo = {
  chainName: string
  chainId: string
  nativeCurrency: {
    name: string
    symbol: string
    decimals: number
  }
  logo?: string
}


//-----------------------------------------------------------------------------------------------//
export default chainInfo


//-----------------------------------------------------------------------------------------------//